project/todo/Todo_IT/src/test/java/io/github/rexmolo/models/AppUserTest.java
package io.github.rexmolo.models;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.NullAndEmptySource;
import org.junit.jupiter.params.provider.ValueSource;

import static org.junit.jupiter.api.Assertions.*;

class AppUserTest {

    @ParameterizedTest
    @NullAndEmptySource
    @ValueSource(strings = {" ", ""})
    void setUserName(String userName) {
        IllegalArgumentException thrown = Assertions.assertThrows(IllegalArgumentException.class, () -> {
            new AppUser(userName, "dddd", AppRole.ROLE_APP_USER);
        });

        Assertions.assertEquals("userName Not allowed to be null", thrown.getMessage());
    }

    @Test
    void getRole() {
    }

    @Test
    void setRole() {
    }

    @Test
    void testEquals() {
    }

    @Test
    void testHashCode() {
    }

    @Test
    void testToString() {
    }
}